:heavy_plus_sign: build: added Multer library for file upload
This change adds the Multer library to the application to allow file uploads. The upload.js file has been created, which defines the necessary settings for the file name after upload and the disk storage provider, using DiskStorage. This defines where files will be saved and how the tmp and upload folders will be managed, including deleting unnecessary files.

Additionally, .gitkeep files have been added to the tmp and upload folders to ensure that these empty folders are kept in Git's version control.

The addition of Multer is important to allow users to upload files to the application. This feature is useful in many cases like allowing users to upload profile pictures or other relevant files. Disk storage provider configuration is also important to ensure that files are managed correctly and that the application is not overloaded with unnecessary files.