:test_tube: test: added avatar image for image upload test
This change adds an avatar image to the application to be used in image upload tests. The image has been added to the tmp/uploads/imageTest folder and will be used as an example during image upload tests.

The addition of this image is important to ensure that the tests are working correctly and that the image upload functionality is being tested properly.