:bug: fix: fixed image upload location
This change fixes the location of uploading images in the application. Previously, images were not being saved in the tmp/uploads folder.

Now the images will be saved in the uploads folder, which is located at the root of the project. To do this, we updated the disk storage provider configuration in the upload.js file to use the correct path for the uploads folder.

This correction is important to ensure that the images are stored correctly and can be easily accessed by the application when needed.