:sparkles: feat: added route, controller, services and repository for avatar upgarde
This change adds avatar update functionality to the app. Added the required route, controller, service, and repository to handle this functionality.

In the users.routes.js route, added the route for updating the avatar, using Multer to handle the image upload. The ProfilePicturesController.js file was also created to control the avatar update and the ProfilePicturesServices.js file to update the avatar image, as well as to delete old images that are no longer in use to avoid server overload. Finally, the ProfilePicturesRepository.js file was created to connect to the database and insert the information in the avatar field.

This functionality is important to allow users to update their avatar images and ensure old images are properly deleted to avoid server overload.