:heavy_check_mark: test: added test for avatares update functionelity
This change adds testing for the update avatars functionality in the app. In ProfilePicturesService.spec.js file upload and unauthorized upload tests have been added.

The ProfileRepositoryInMemory.js file was also created to ensure that no data is inserted into the database during testing.

These tests are important to ensure that the avatar update functionality is working correctly and to identify possible problems or bugs in the application. This increases application reliability and helps prevent problems for end users.