:heavy_plus_sign: build: added jsonwebtoken library
In this change, the jsonwebtoken library was added to allow the generation and validation of JWT tokens (JSON Web Tokens). These tokens will be used to authenticate and authorize users in different application routes.

The inclusion of the jsonwebtoken library is important to ensure the security of the application's information and to prevent unauthorized access to protected resources and data. Furthermore, the library simplifies the generation and validation of JWT tokens, making the process easier and more reliable.