:sparkles: feat: added sessions files
In this change, four files were added for creating user authorization sessions in the application. The sessionsRoutes.js, SessionsController.js, SessionsService.js and SessionsRepository.js files are responsible for creating routes, controlling sessions, handling session data and connecting to the database.

The inclusion of these files is essential to ensure application security and protect user data. With the implementation of these sessions, it is possible to authenticate and authorize user access to the application in a safe and controlled way, in addition to preventing unauthorized access and protecting application data.