:heavy_check_mark: test: added files for session tests
In this change, two files were added to test the functionality of sessions in the application. The SessionsService.spec.js and SessionsRepositoryInMemory.js files are responsible for simulating the connection to the database and carrying out user authentication and authorization tests in the application.

By implementing these tests, it is possible to ensure that session functionalities are working correctly and that user data is being treated in a safe and controlled manner in the application.