:recycle: refactor: change in capturing logged in user ID
Previously, capturing the logged in user's id was done using the constant { id } = request.query. However, with the addition of user authentication and session creation, this process needed to be modified. Now, the logged in user id is captured through the constant user_id = request.user.id, ensuring greater security and efficiency in the system.