:sparkles: feat: adding files related to cymbal functionality
This commit includes the addition of new files related to system platter functionality. The files dishes.routes.js, DishesController.js, DishesServices.js and DishesRepository.js were added, which are responsible for implementing the business logic, managing the routes, performing the operations in the database and providing the necessary services for manipulation of dish data.