:ok_hand: style: default quotes to ' throughout project
A general correction in the project's writing was made to standardize the use of single quotes instead of double quotes. In many files, double quotes were inadvertently used, which created inconsistencies in the code and made it difficult to read. To address this issue, we scanned the entire project to replace double quotes with single quotes. This change is important because it follows the project's writing conventions and makes the code more readable and easier to understand.