:sparkles: feat: adding all routes
This commit adds all routes available in the application to the index.js file in the routes folder. This ensures that all routes are available in a single file, making the code easier to navigate and maintain. With this updated file, the routes will be available to the entire application and can be easily accessed by other modules. This change provides a clearer organization of the code and makes it easier to add new routes in the future.