:heavy_plus_sign: build: addition of Cors
This update adds Cors to the project. Cors is a technique used to allow a server to receive requests from other domains, allowing web applications to interact with different servers.

With CORS, it will be possible to make requests to external APIs and receive data to be displayed in the application, even if these APIs are hosted on different servers.