:bricks: ci: favorites table
This update added a new migration to the Favorite Items functionality, which created a table named 'favorites' in the database with fields such as auto-incremented 'id'. In addition, this table is connected to the user table by the 'user_id' and to the dish table by the 'dish_id'. This will allow users to add dishes to their favorites and access them easily on the platform.