:sparkles: feat: favorites
This update added several files and functionality related to creating routes, controllers, services and repositories for the bookmarking functionality.

The 'favoritesRoutes.js' file was created to define the routes needed for create, display, list and delete favorite items operations. The 'FavoritesController.js' file is responsible for receiving information from the front-end via request and sending the appropriate response via response.

The 'FavoritesService.js' file is responsible for the business rules and for executing the requested operations, such as creating, displaying, listing and deleting favorite items. And, finally, the 'FavoritesRepository.js' file is responsible for connecting to the database and performing the CRUD operations (create, read, update and delete) for the favorites table, using the user ID and Dish ID.