:bricks: ci: routes
This update consisted of including favorite routes in the 'index' file in the server's 'routes' folder, which allowed them to become available to the application's front-end. Favorite routes include the basic CRUD operations (create, read, update, and delete) to manage a user's list of favorite dishes, allowing the user to add or remove dishes from their list.