:sparkles: feat: purchase affection table
This update aims to create a table called 'carts' in the project database. This table will have a relationship with the 'users' and 'dishes' tables through the 'id' field, and will have the functionality to store the quantity of each product added to the cart, through the 'quantity' column. The 'onDelete' option was also configured to ensure data integrity, allowing that, in case a user or a plate is deleted, the corresponding information in the 'carts' table is also removed. This ensures that the database is kept organized and consistent.