:sparkles: feat: CRUD shopping affection
In this update, the 'carts.routes.js' file was created, which has the routes 'create', 'update', 'index' and 'delete'. The 'CartsController.js' file was also created, which contains the controller class responsible for receiving route requests. In addition, the 'CartsServices' file was created, which handles the system logic through requests sent by the controller. Finally, the 'CartsRepository.js' file was created, being responsible for the database connection requested by the service. These additions allow operations related to shopping carts to be managed efficiently and neatly.