:ok_hand: style: ordes controller
This update changes the way of sending information that used to be done by returning an object with an 'orders' key that contained the value of the 'orders' variable. With the update it was changed to directly send the value of 'orders' without an object with a specific key. This change was made to the logic of the code responsible for sending the HTTP response to the client after a request. With this change, the process of sending information is simpler and more straightforward, without the need to add a key to the returned object.