:bug: fix: order creation and indexing
This update identified an error in creating and indexing orders.

When searching for the dish_id, it was looking for the id of the cart previously created, which was incorrect. By changing it to search for item.dish_id created with the cart map, the creation error was fixed.

When verifying if a user was an administrator, the 'checkUserIsAdmin' if statement was used, which always returned true because it was always searching for an existing user. So, we changed this comparison to 'checkUserIsAdmin.isAdmin === 1'. This way, it only returns true if the user is actually an administrator.