:sparkles: feat: carts
This update added the carts route to the index.js file in the routes folder, allowing the previously created shopping cart functionality to be used in the application. By inserting this route, it will be possible to manage the items added to the cart by the users, allowing the visualization of the selected products, quantity, price, in addition to the finalization of the purchase.