:sparkles: feat: users repository
In the update, there was a modification in the way of searching data in the usersRepository, more specifically in the functions findByEmail, show and update. Previously, these functions took a single argument, but now they take an object. This change can be beneficial for the code, since the object can contain several pieces of information at the same time and facilitate the search and manipulation of the data.