:recycle: refator: favorites repository
In this update, an optimization was made in the database. Previously, an SQL query was executed to select unnecessary information that would not be used in the frontend and that was not consumed by the API. Now, this information has been removed from the query so that the component rendering process is faster and more efficient. With this, the application becomes more agile and improves the user experience.