:sparkles: feat: script and dependencies
In this update PM2 was added to the project with the previous update, which created the 'ecosystem.config.js' file, PM2 will be able to run the API 24 hours a day, 7 days a week. The 'pm2-runtime start ecosystem.config.js --env production' script has been added to run the 'ecosystem.config.js' file and the settings it contains. This update ensures that the API runs continuously even in case of server crashes or restarts.