:sparkles: feat configs
This update made a number of design changes. First, the direct dependency on the database was removed, leaving Knex to manage it. Next, the uploadConfig procedure was added to the server, so that it is possible to configure the default image search page. Finally, CORS was added to the project to allow access to resources in other domains.