:books: docs: README
This update added the 'README' file to the project, containing initial information about the project, the technologies used and how to run it in a local environment. This file is essential for guiding other developers who might collaborate or review the code. In addition, the readme file will be updated with more relevant information, such as the link to the final project and instructions for use.