:rocket: image
This update adds the images of the dishes and the standard users in the project's images folder, allowing them to be used in the final version of the application. Dish images will be displayed on the menu page, while default user images will be used on the profile page and header. These images serve to make the application's interface more attractive and informative for users.