:goal_net: database
This update removed the default database previously used at the beginning of the deploy and inserted a new one with little information, as the old one was showing incorrect behavior towards the end user. Tests will be carried out with this new database to diagnose the error and continue the project.