:sparkles: feat: users
This update changes the access path to the Knex module, plus it adds a check to verify that the database has already been created before creating the admin user. This is done to ensure the database is available before creating the user with administrator privileges. This check is important because if the database has not yet been created, trying to create the admin user would result in an error. With this change, the system becomes more robust and avoids possible failures.