:bug: fix: password
In this update, the 'hash' function from the 'bcryptjs' library has been added. The admin password was previously being stored in plaintext, the change was necessary because the cleartext password was causing errors when using the 'compare' function of 'bcryptjs', which is used to verify the match of the encrypted password with the password provided by the user.