:test_tube: test: table
In this update, an SQL command was added to query which tables are being created in the database, using the syntax 'SELECT name FROM sqlite_master WHERE type='table''. This command returns a list with the names of all the tables present in the SQLite type database that are currently being created.