:sparkles: feat: array
In this update, a change was made to the way data is inserted from an array. The use of the 'map' method was implemented together with the destructuring of the variables that you want to insert, making the code simpler and easier to understand. This change aims to improve the maintainability and readability of the code.