:sparkles: feat: dishes
In this update, the way of inserting data in the table of ingredients in the database was changed. Previously, insertion was done using a loop in the map format, while now insertion is done using the Promise.all() function in conjunction with the map method. This way, you can insert multiple items at the same time asynchronously and wait for all insertions to complete before proceeding with the code.