:bug: fix: SQLITE_CONSTRAINT: FOREIGN KEY constraint failed
In this update, a workaround has been implemented to fix the bug reported in issue #1. The [id] parameter was destructured during the creation of a new request, ensuring that data insertion does not use the foreign key incorrectly and thus avoids the restriction error. This change aims to ensure that the insertion of orders is carried out correctly, respecting the restrictions and relationships defined in the database.