:ok_hand: style: database
In this update, the direct database insertion logic present in the server.js file has been removed. Instead, the project returns Knex to carry out this action.
This change provides a more organized and readable code, in addition to improving the maintainability and scalability of the project. Knex becomes the intermediary between the application and the database, ensuring an abstraction layer and simplifying interactions with the database.